Команда save

+++ src/main/java/ru/itmo/ArsikAndEva/storage/AllData.java
@@ -5,6 +5,6 @@ import ru.itmo.ArsikAndEva.model.*;
 import java.io.Serializable;
 import java.util.HashMap;
 
-public record AllData(HashMap<Integer, Booking> bookings, HashMap<Integer, Instrument> instruments,
-                      HashMap<Integer, Checkout> checkouts) implements Serializable {
+public record AllData(HashMap<Long, Booking> bookings, HashMap<Long, Instrument> instruments,
+                      HashMap<Long, Checkout> checkouts) implements Serializable {
 }
